Add SDK functions
